requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.2
catboost==1.2.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
contourpy==1.2.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.4.3
cycler==0.12.1
dataclasses-json==0.6.4
db-dtypes==1.2.0
debugpy==1.8.0
decorator==4.4.2
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.8
distro==1.9.0
Django==5.0.3
django-environ==0.11.2
django-rest-swagger==2.2.0
djangorestframework==3.15.1
docstring_parser==0.16
docutils==0.20.1
python-dotenv
elastic-transport==8.12.0
elasticsearch==8.12.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
ez-setup==0.9
faiss-cpu==1.8.0
Faker==24.1.0
fastjsonschema==2.19.1
favicon==0.7.0
ffmpeg-python==0.2.0
filelock==3.13.1
fonttools==4.49.0
fqdn==1.5.1
frozenlist==1.4.1
funcsigs==1.0.2
future==1.0.0
fuzzywuzzy==0.18.0
gcloud==0.18.3
gensim==4.3.2
gitdb==4.0.11
GitPython==3.1.42
google-ai-generativelanguage==0.4.0
google-api-core==2.16.1
google-api-python-client==2.127.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-aiplatform==1.49.0
google-cloud-bigquery==3.17.1
google-cloud-bigquery-storage==2.24.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-speech==2.26.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-generativeai==0.3.2
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
GoogleBard==2.1.0
graphviz==0.20.1
grpc-google-iam-v1==0.13.0
grpcio==1.60.1
grpcio-status==1.60.1
gspread==5.12.4
gspread-pandas==3.3.0
h11==0.14.0
htbuilder==0.6.2
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
idna==3.6
imageio==2.34.1
imageio-ffmpeg==0.4.9
importlib-metadata==7.0.1
iniconfig==2.0.0
ipykernel==6.29.0
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
itypes==1.2.0
jaraco.classes==3.3.1
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-highlight-selected-word==0.2.0
jupyter-lsp==2.2.2
jupyter-nbextensions-configurator==0.6.3
jupyter_client==8.6.0
jupyter_contrib_core==0.4.2
jupyter_contrib_nbextensions==0.7.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.0.12
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keyring==24.3.1
kiwisolver==1.4.5
lancedb==0.6.13
langchain==0.1.6
langchain-community==0.0.19
langchain-core==0.1.22
langchain-experimental==0.0.50
langchain-google-genai==0.0.9
langdetect==1.0.9
langsmith==0.0.87
Levenshtein==0.25.0
lxml==5.1.0
Markdown==3.5.2
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.4
marshmallow==3.20.2
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.2.0
moviepy==1.0.3
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.6.0
nh3==0.2.15
nltk==3.8.1
notebook==7.0.7
notebook_shim==0.2.3
numpy==1.26.3
oauth2client==4.1.3
oauthlib==3.2.2
openai==0.28.0
openapi-codec==1.3.2
openpyxl==3.1.2
overrides==7.7.0
packaging==23.2
pandas==1.5.3
pandas-gbq==0.20.0
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pkginfo==1.10.0
platformdirs==4.2.0
plotly==5.19.0
pluggy==1.4.0
proglog==0.1.10
prometheus-client==0.19.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.0
pydantic_core==2.16.1
pydata-google-auth==1.8.2
pydeck==0.8.1b0
Pygments==2.17.2
pylance==0.10.12
pymdown-extensions==10.7.1
pyparsing==3.1.1
pypdf==4.1.0
PyPDF2==3.0.1
pytest==8.1.0
pytest-cov==2.12.1
pythainlp==5.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-Levenshtein==0.25.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
rapidfuzz==3.6.2
ratelimiter==1.2.0.post0
readme_renderer==43.0
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
retry==0.9.2
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
scikit-learn==1.4.1.post1
scikit-surprise==1.1.3
scipy==1.12.0
seaborn==0.11.2
semver==3.0.2
Send2Trash==1.8.2
shapely==2.0.4
simplejson==3.19.2
six==1.16.0
smart-open==7.0.1
smmap==5.0.1
sniffio==1.3.0
socksio==1.0.0
soupsieve==2.5
SQLAlchemy==2.0.26
sqlparse==0.4.4
st-annotated-text==4.0.1
stack-data==0.6.3
streamlit==1.32.1
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.0
streamlit-embedcode==0.1.2
streamlit-extras==0.4.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.3
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
StrEnum==0.4.15
surprise==0.1
tabulate==0.9.0
tenacity==8.2.3
terminado==0.18.0
threadpoolctl==3.3.0
tiktoken==0.6.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
twine==4.0.2
types-python-dateutil==2.8.19.20240106
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.0
vertexai==1.49.0
virtualenv==20.25.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
widgetsnbextension==4.0.9
wrapt==1.16.0
xlrd==2.0.1
yarl==1.9.4
zipp==3.17.0
